[FEATURE-12] Refactor Application project
