+++ src/Host/Startup.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
-using YAGO.Services.WeatherForecasts;
+using YAGO.WebProjectTemplate.Application;
 
 namespace YAGO.WebProjectTemplate.Web
 {
